requirements.txt
aiodns
aiohttp
async-timeout
fire
progressbar2
requests
